+++ src/group/group_builder.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use crate::metadata::{v3::GroupMetadataV3, GroupMetadata};
+use crate::metadata::{v3::GroupMetadataV3, AdditionalFields, GroupMetadata};
 
 use super::{Group, GroupCreateError};
 
@@ -43,13 +43,10 @@ impl GroupBuilder {
     ///
     /// Note that array metadata must not contain any additional fields, unless they are annotated with `"must_understand": false`.
     /// `zarrs` will error when opening an array with additional fields without this annotation.
-    pub fn additional_fields(
-        &mut self,
-        additional_fields: serde_json::Map<String, serde_json::Value>,
-    ) -> &mut Self {
+    pub fn additional_fields(&mut self, additional_fields: AdditionalFields) -> &mut Self {
         match &mut self.metadata {
-            GroupMetadata::V3(metadata) => metadata.additional_fields = additional_fields.into(),
-            GroupMetadata::V2(metadata) => metadata.additional_fields = additional_fields.into(),
+            GroupMetadata::V3(metadata) => metadata.additional_fields = additional_fields,
+            GroupMetadata::V2(metadata) => metadata.additional_fields = additional_fields,
         };
         self
     }
@@ -82,7 +79,7 @@ mod tests {
         attributes.insert("key".to_string(), "value".into());
         builder.attributes(attributes.clone());
 
-        let mut additional_fields = serde_json::Map::new();
+        let mut additional_fields = AdditionalFields::new();
         let additional_field = serde_json::Map::new();
         additional_fields.insert("key".to_string(), additional_field.into());
         builder.additional_fields(additional_fields.clone());